requirements.txt
fastapi~=0.70.0
SQLAlchemy~=1.4.25
requests~=2.26.0
python-dotenv~=0.19.1
uvicorn~=0.15.0
pydantic~=1.8.2
psycopg2~=2.9.1
gunicorn~=20.0.1